refactor: rename dataframe module to players_data
