+++ server_manager/src/set_mode.py
@@ -5,7 +5,7 @@ import logging
 import sys
 from typing import List
 
-from .dataframe import get_mode, get_username, get_uuid
+from .players_data import get_mode, get_username, get_uuid
 from .exceptions import InvalidPlayerDataStateError, InvalidServerStateError
 from .player import Player
 from .properties_manager import get_server_mode, get_server_path, set_server_mode
